refactor(ingest): rename flocks/syslog → flocks/ingest/syslog
Group data-ingestion triggers under a shared flocks/ingest/ namespace so
that future connectors (kafka, webhook, …) live alongside syslog as
sibling sub-packages rather than top-level packages.

- Move flocks/syslog/* → flocks/ingest/syslog/*
- Add flocks/ingest/__init__.py (empty package marker)
- Update all internal imports (flocks.syslog → flocks.ingest.syslog)
  in listener.py, manager.py, __init__.py, server/app.py and
  server/routes/workflow.py
