+++ flocks/ingest/syslog/__init__.py
@@ -1,6 +1,6 @@
 """Syslog ingestion for workflow triggers (UDP/TCP listeners)."""
 
-from flocks.syslog.constants import WORKFLOW_SYSLOG_CONFIG_PREFIX
-from flocks.syslog.manager import SyslogManager, default_manager
+from flocks.ingest.syslog.constants import WORKFLOW_SYSLOG_CONFIG_PREFIX
+from flocks.ingest.syslog.manager import SyslogManager, default_manager
 
 __all__ = ["SyslogManager", "default_manager", "WORKFLOW_SYSLOG_CONFIG_PREFIX"]
