+++ flocks/ingest/syslog/constants.py
