+++ flocks/ingest/syslog/parser.py
